19364 - moved helper method to faa helper
